Format and Fluidify Definition
Changes one or more HTML Definition Lists to show it’s groups of Dt and Dd elements inside floating Divs. The wrapping Divs are created and the Flex CSS of the Dl is set initially.
It is also easy to format the elements inside the Dl further by setting style or class by tag name.